+++ config/application.rb
@@ -1,6 +1,7 @@
 require_relative "boot"
 
 require "rails/all"
+require_relative "../lib/app_host"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.